AlchemyAPI key added
This is our private key (that is now publicly available…) required when
using the API.

For now, it resides in the directory that the supplied test cases are
to be run from.

e.g. run the following after using ant to build from the testdir

‘java -jar ../dist/AlchemyAPI-Concept-Test.jar’

will give a flavour of the default output.